# api-gateway/internal/app/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a restaurant reservation system.",
        "title": "Artisan Connect",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9006",
    "basePath": "/",
    "paths": {
...
        "/report/chosen/{start_at}/{end_at}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetFromTill  expenses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "GetFromTill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start_at",
                        "name": "start_at",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "end_at",
                        "name": "end_at",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.FromTillResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/expenses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetExpenses  expenses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "GetExpenses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.IncomeExpenseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/spending": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetSpending  speding",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "GetSpending",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ListSpendingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.FromTillResponse": {
            "type": "object",
            "properties": {
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "entity.IncomeExpenseResponse": {
            "type": "object",
            "properties": {
                "net_savings": {
                    "type": "number"
                },
                "total_expenses": {
                    "type": "number"
                },
                "total_income": {
                    "type": "number"
                }
            }
        },
...
        "entity.ListSpendingResponse": {
            "type": "object",
            "properties": {
                "spent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SpendingResponse"
                    }
                }
            }
        },
...
        "entity.SpendingResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "total_spent": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}